backend/requirements.txt
boto3==1.4.8
botocore==1.8.1
certifi==2017.7.27.1
chardet==3.0.4
decorator==4.1.2
Django==1.11.7
django-cors-middleware==1.3.1
django-extensions==1.9.7
django-filter==1.1.0
django-oauth-toolkit==1.0.0
django-storages==1.6.5
djangorestframework==3.7.1
docutils==0.14
drf-writable-nested==0.3.1
factory-boy==2.9.2
Faker==0.8.6
gunicorn==19.7.1
idna==2.6
ipython==6.2.1
ipython-genutils==0.2.0
jedi==0.11.0
jmespath==0.9.3
Markdown==2.6.9
oauthlib==2.0.6
olefile==0.44
parso==0.1.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.3.0
pprint==0.1
prompt-toolkit==1.0.15
psycopg2==2.7.3.2
ptyprocess==0.5.2
Pygments==2.2.0
python-dateutil==2.6.1
python-http-client==3.0.0
pytz==2017.3
requests==2.18.4
s3transfer==0.1.11
sendgrid==5.3.0
simplegeneric==0.8.1
six==1.11.0
tblib==1.3.2
text-unidecode==1.0
traitlets==4.3.2
typing==3.6.2
urllib3==1.22
wcwidth==0.1.7
